requirements.txt
asgiref==3.5.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.4
django-ckeditor==6.4.0
django-etc==1.3.1
django-hitcount==1.3.5
django-js-asset==2.0.0
gunicorn==20.1.0
idna==3.3
Pillow==9.1.0
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
